move all CofkCollect* table to uploader
